added: CreateBalanceAndBalanceAuditTest, SendCreateAccountNotificationTest

+++ src/main/java/faang/school/accountservice/service/request_task/handler/impl/create_account/CreateBalanceAndBalanceAudit.java
@@ -11,6 +11,7 @@ import faang.school.accountservice.entity.RequestTask;
 import faang.school.accountservice.enums.request_task.RequestTaskStatus;
 import faang.school.accountservice.enums.request_task.RequestTaskType;
 import faang.school.accountservice.exception.JsonMappingException;
+import faang.school.accountservice.repository.AccountRepository;
 import faang.school.accountservice.service.BalanceAuditService;
 import faang.school.accountservice.service.BalanceService;
 import faang.school.accountservice.service.request.RequestService;
@@ -35,6 +36,7 @@ public class CreateBalanceAndBalanceAudit implements RequestTaskHandler {
     private final BalanceService balanceService;
     private final RequestService requestService;
     private final BalanceAuditService balanceAuditService;
+    private final AccountRepository accountRepository;
 
     private final CheckAccountsQuantity checkAccountsQuantity;
     private final CreateAccount createAccount;
@@ -48,7 +50,10 @@ public class CreateBalanceAndBalanceAudit implements RequestTaskHandler {
     @Override
     public void execute(Request request) {
         try {
-            Account account = mapAccount(request);
+            Long accountId = Long.valueOf(request.getContext());
+            Account account = accountRepository.findById(accountId)
+                    .orElseThrow(() -> new EntityNotFoundException("Account with Id: " +
+                            accountId + " not found"));
             BalanceDto balanceDto = balanceService.createBalance(account);
             List<BalanceAudit> audits = account.getBalanceAudits();
             List<Long> auditsIds = audits.stream()
@@ -62,8 +67,8 @@ public class CreateBalanceAndBalanceAudit implements RequestTaskHandler {
             String rollbackContext = mapRollbackTaskDtoToString(context);
             setRequestTaskStatusAndContext(request, RequestTaskStatus.DONE, rollbackContext);
             requestService.updateRequest(request);
-            log.info("Finished processing request task with type: {}",
-                    RequestTaskType.WRITE_INTO_BALANCE_BALANCE_AUDIT);
+            log.info("Finished processing request task with type: {}, id: {}",
+                    RequestTaskType.WRITE_INTO_BALANCE_BALANCE_AUDIT, request.getIdempotentToken());
         } catch (OptimisticLockingFailureException e) {
             log.error("Optimistic locking failed after 3 retries for request with id: {}. " +
                     "Executing rollback.", request.getIdempotentToken(), e);
@@ -71,8 +76,9 @@ public class CreateBalanceAndBalanceAudit implements RequestTaskHandler {
             throw e;
         } catch (Exception e) {
             log.error("Unexpected error occurred during execution request with id: {}. " +
-                    "Executing rollback.",request.getIdempotentToken(), e);
+                    "Executing rollback.", request.getIdempotentToken(), e);
             rollback(request);
+            throw e;
         }
     }
 
@@ -81,24 +87,29 @@ public class CreateBalanceAndBalanceAudit implements RequestTaskHandler {
         return 3;
     }
 
+    @Transactional
     @Override
     public void rollback(Request request) {
         RequestTask requestTask = request.getRequestTasks().stream()
                 .filter(task -> task.getHandler().equals(RequestTaskType.WRITE_INTO_BALANCE_BALANCE_AUDIT))
                 .findFirst().orElseThrow(() -> new EntityNotFoundException("No request task found"));
 
-        RollbackTaskCrateBalanceDto dto = mapRollbackTaskDto(requestTask.getRollbackContext());
-        balanceService.deleteBalance(dto.getBalanceId());
-        dto.getBalanceAuditIds().forEach(balanceAuditService::deleteAudit);
-        setRequestTaskStatusAndContext(request, RequestTaskStatus.AWAITING, null);
+        if (requestTask.getRollbackContext() != null) {
+            RollbackTaskCrateBalanceDto dto = mapRollbackTaskDto(requestTask.getRollbackContext());
+            balanceService.deleteBalance(dto.getBalanceId());
+            dto.getBalanceAuditIds().forEach(balanceAuditService::deleteAudit);
+        }
 
         List<RequestTask> tasks = request.getRequestTasks().stream()
                 .filter(task -> task.getStatus() == RequestTaskStatus.DONE).toList();
+
         if (!tasks.isEmpty()) {
+            setRequestTasksStatus(request, RequestTaskStatus.AWAITING);
             checkAccountsQuantity.rollback(request);
             createAccount.rollback(request);
         }
-        log.info("Request task with type: {} rollback",RequestTaskType.WRITE_INTO_BALANCE_BALANCE_AUDIT);
+        log.info("Request task with id: {}, type: {} rollback",
+                request.getIdempotentToken(), RequestTaskType.WRITE_INTO_BALANCE_BALANCE_AUDIT);
     }
 
     private String mapRollbackTaskDtoToString(RollbackTaskCrateBalanceDto dto) {
@@ -121,21 +132,16 @@ public class CreateBalanceAndBalanceAudit implements RequestTaskHandler {
         return dto;
     }
 
-    private Account mapAccount(Request request) {
-        Account account;
-        try {
-            account = objectMapper.readValue(request.getContext(), Account.class);
-        } catch (JsonProcessingException e) {
-            throw new JsonMappingException(e.getMessage());
-        }
-        return account;
+    private void setRequestTasksStatus(Request request, RequestTaskStatus status) {
+        request.getRequestTasks()
+                .forEach(requestTask -> requestTask.setStatus(status));
     }
 
     private void setRequestTaskStatusAndContext(
             Request request, RequestTaskStatus requestTaskStatus, String taskContext) {
         request.getRequestTasks().stream()
                 .filter(requestTask -> requestTask.getHandler().
-                        equals(RequestTaskType.WRITE_INTO_ACCOUNT))
+                        equals(RequestTaskType.WRITE_INTO_BALANCE_BALANCE_AUDIT))
                 .forEach(requestTask -> {
                     requestTask.setStatus(requestTaskStatus);
                     requestTask.setRollbackContext(taskContext);
